bot/requirements.txt
# requirements.txt

# Personal access token grants access to repositories in dpsh organization.
codec @ git+https://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@github.com/dpshai/codec.git

# FastAPI dependencies.
aiosmtplib
fastapi[all]
mangum
pyjwt
uvicorn

# Database dependencies.
tortoise-orm[asyncpg]

# Testing dependencies.
pytest
